fix(build_library): use https://public.update.core-os.net now
We added a new https certificate on the new update service and changed
the hostname to be consistent with all of the other endpoints. Update
the new images to use this.

The old URL http://public.roller.core-os.net will remain working until
all of the old client have been updated.