fix(core_roller_upload): Use new payload names and upload path.
As of 4e85b172 the build system generates the update payload using the
coreos_production_update prefix, so use that instead. Use the release
version in the storage path instead of a disk image MD5; the version is
sufficiently unique and the no need for the obfuscation.

Drop the disk image copy in the download storage that I tacked in, that
will go into a new core_promote script that I have written but not
tested yet (need an alpha release to promote!!).