build_packages: always disable --rebuild-if-unbuilt on first installs
I'm not sure at what point the rebuild option started triggering so many
rebuilds but it always does these days. On the first install any builds
it triggers are probably pointless anyway, making it take way way to
long to get started with a fresh SDK or after resetting the board root.