catalyst: hacky fix for amd64 toolchain builds
The mpc upgrade is breaking the board toolchain build because the
two-stage build process is improperly kicking in, the wrong libraries
get linked, sysroot ignored, and other unpleasentries.

This should work around the issue pending fixing the gcc ebuild or
whatever is allowing the two-stage build to kick in for cross-compiled
builds.