build_torcx_store: add a command for building torcx images
This creates new torcx stores under src/build/torcx, similar to the
layout of src/build/images.  The script contains a list of packages
to build so there is a deterministic set that gets branched with
this repository.  For development, a different list can also be
given on the command line.

Images are created by building only the direct dependencies of
meta-packages under the app-torcx category.  They use the board
root as the sysroot for cross-compiling.  The base layout of the
installation root is structured so everything is under either /bin
or /lib for simplicity.  Any systemd units are rewritten so they
depend on a successful torcx apply, and they inject the appropriate
torcx runtime directory into the PATH.

When ELF binaries are compiled, they are given an RPATH value of
/ORIGIN/../lib which gets rewritten to $ORIGIN/../lib.  The final
value enables automatic dynamic linking with packaged libraries.
The intermediate value is to avoid having to escape the $ through
the various build system layers (which are different for each
package) and it pretends to be an absolute path to silence security
warnings.