build_packages: also build all torcx images by default
The idea is that once the installed board packages have changed,
the sysroot that was used to build any existing torcx packages may
no longer be compatible.  It will therefore run build_torcx_store,
passing relevant options to it, to build a new torcx store of
images with the current sysroot.

This adds --skip_torcx_store to use the previous behavior.