update_ebuilds: replace cvs with git
And default to git instead of rsync.

git has no rate limiting and will generally be quicker after the first
run.

This does leave a bit of extra data in your local portage-stable `.git`
directory, but it doesn't seem unreasonable to me.

Note: this means we lose the "ChangeLog" file. In the rsync
repositories, that file has been generated by egencache, but the git
repository never has it checked in.