build_library/torcx_manifest: omit newlines
Omit the newline for each function that returns only a single value.

This does not change each function that returns multiple values as those
are expected, by the caller, to be newline delimited values.

Before this change, the consumer of 'local_store_path' specifically
ended up with bad behavior of including a newline in the on-disk
filename.

Fortunately, the symlink included that newline as well, so it didn't
*break* things, but it did mean the filepath differed from the one
declared in the manifest slightly.