catalyst: Disable ccache
It provides no value when it works, and it's randomly causing
failures to build toolchains due to permissions problems after
certain releases.  This also requires taking it out of FEATURES in
the portage profile (which is the SDK profile by default).

Test Jenkins runs of SDK and toolchains jobs both ran in the same
time as with ccache enabled.