transfer.sh: new ready/roll command series
the ready command prepares the specified groups for the rollout - it
sets the channel to the new version and updates the update percent to
0%. this allows machines to force update, but updates don't get rolled
out automatically.

the roll command configures the rollout strategy. right now it creates a
linear rollout, going from 0% to 100% in the given number of hours for
each group. it uses the default frame size of 60 seconds (it sets it to
be explicit).